strdup2

+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: yuonishi <yuonishi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/19 16:21:05 by yuonishi          #+#    #+#             */
-/*   Updated: 2025/11/01 20:19:36 by yuonishi         ###   ########.fr       */
+/*   Updated: 2025/11/02 21:44:58 by yuonishi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
